requirements.txt
opencv-python==4.11.0.86
matplotlib==3.10.1
numpy==1.25.0
Pillow==11.1.0
torch==2.6.0+cpu
torchvision==0.21.0+cpu
albumentations==2.0.5
pyyaml==6.0.2
tqdm==4.67.1
seaborn==0.13.2
scikit-learn==1.6.1
ultralytics==8.3.105
